Redesigned
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@24016 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ scripts/attach_project.sh
@@ -5,7 +5,6 @@
 
 #-----------------------------------------------------------------------------
 #set -xv
-#set -x
 
 #-----------------------------------------------------------------------------
 # defaults
@@ -32,12 +31,11 @@ SYNOPSIS:
 ARGUMENTS:
   proj -- mandatory name of proj.attach file in scripts/projects or
           scripts/internal/projects folder
-  -n   -- do not actually do anything, just report
+  -n   -- do not change anything, just report
 EOF
     test -z "$1"  ||  echo ERROR: $1 1>&2
     exit 1
 }
-
 
 #-----------------------------------------------------------------------------
 # analyze script arguments
@@ -49,7 +47,7 @@ fi
 for cmd_arg in "$@"; do
   case "$cmd_arg" in
     -n ) do_update="no" ;;
-    *  ) if -n "$proj"
+    *  ) if test -n "$proj"; then
            Usage "Only one project file is allowed"
          else
            proj="$cmd_arg"
@@ -58,41 +56,47 @@ for cmd_arg in "$@"; do
   esac
 done
 
-
-
 #-----------------------------------------------------------------------------
-# check existence of files and folders
+# check existence of input file
 
-proj_file="$proj"
-if test ! -f "$proj_file" ; then
-  proj_file="$script_dir/projects/$proj.attach"
-  if test ! -f $proj_file ; then
-    proj_file="$script_dir/internal/projects/$proj.attach"
+prjfile="$proj"
+if test ! -f "$prjfile" ; then
+  prjfile="$script_dir/projects/$proj.attach"
+  if test ! -f "$prjfile" ; then
+    prjfile="$script_dir/internal/projects/$proj.attach"
   fi
 fi
-test -f $proj_file  ||  Usage "Project file '$proj.attach' not found"
+test -f "$prjfile"  ||  Usage "Project file '$proj.attach' not found"
 
 
 #-----------------------------------------------------------------------------
-
-cat "$proj_file" | while read dir cvs_loc
+# check existence and CVS status of folders
+cat "$prjfile" | while read src_loc cvs_loc
 do
-  cd "$initial_dir"
-  if test ! -d "$dir" ; then
-    COMMON_Error "$dir is not found"
-  fi
-  dir_name=`dirname $dir`
-  base_name=`basename $dir`
+  test ! -d "$src_loc" && COMMON_Error "$src_loc is not found"
+done
+
+# replace folders
+cat "$prjfile" | while read src_loc cvs_loc
+do
+  dir_name=`dirname $src_loc`
+  base_name=`basename $src_loc`
   echo "=================================================================="
+  COMMON_Exec cd "$initial_dir/$src_loc"
+  files=`ls | sed -e '/CVS/d'`
+  if test -n "$files"; then
+    echo "WARNING: $src_loc is not empty -- SKIPPING"
+    continue
+  fi
+  COMMON_Exec cd "$initial_dir/$dir_name"
   if test "$do_update" = "yes" ; then
-    COMMON_Exec cd $dir_name
-    echo "removing existing $dir"
+    echo "Removing empty $src_loc"
     COMMON_Exec rm -r "$base_name"
     echo "------------------------------------------------------------------"
-    echo "CVS checkout $cvs_loc into $dir"
+    echo "CVS checkout $cvs_loc into $src_loc"
     COMMON_Exec cvs co -d "$base_name" "$cvs_loc" > /dev/null
   else
-    echo "$dir to be replaced by files from CVS in $cvs_loc"
+    echo "$src_loc to be replaced by files from CVS in $cvs_loc"
   fi
 done
